Remove Specimen, which has been deprecated for a while, and introduce SampleHelper for the one method still used.
