+++ .gitignore
@@ -18,7 +18,7 @@ docker/ingest-api/BUILD
 BUILD
 
 # Ignore the docker mounted BUILD and VERSION
-src/BUILD
+BUILD
 src/VERSION
 
 # Ignore __pycache__ directories in any path
